+++ src/connectionMatrix/view.cpp
@@ -91,6 +91,7 @@ View::View(QWidget* parent)
 	settings->registerSettingObserver(settings::ConnectionMatrix_ShowMediaLockedDot.name, this);
 	settings->registerSettingObserver(settings::ConnectionMatrix_AllowCRFAudioConnection.name, this);
 	settings->registerSettingObserver(settings::ConnectionMatrix_CollapsedByDefault.name, this);
+	settings->registerSettingObserver(settings::ConnectionMatrix_ShowEntitySummary.name, this);
 	settings->registerSettingObserver(settings::General_ThemeColorIndex.name, this);
 
 	// react on connection completed signals to show error messages.
@@ -109,6 +110,7 @@ View::~View()
 	settings->unregisterSettingObserver(settings::ConnectionMatrix_ShowMediaLockedDot.name, this);
 	settings->unregisterSettingObserver(settings::ConnectionMatrix_AllowCRFAudioConnection.name, this);
 	settings->unregisterSettingObserver(settings::ConnectionMatrix_CollapsedByDefault.name, this);
+	settings->unregisterSettingObserver(settings::ConnectionMatrix_ShowEntitySummary.name, this);
 	settings->unregisterSettingObserver(settings::General_ThemeColorIndex.name, this);
 }
 
@@ -613,6 +615,13 @@ void View::onSettingChanged(settings::SettingsManager::Setting const& name, QVar
 
 		// Manually force a model refresh of the headers
 		_model->forceRefreshHeaders();
+	}
+	else if (name == settings::ConnectionMatrix_ShowEntitySummary.name)
+	{
+		auto const drawSummary = value.toBool();
+		_itemDelegate->setDrawEntitySummary(drawSummary);
+
+		forceFilter();
 	}
 	else if (name == settings::General_ThemeColorIndex.name)
 	{