+++ src/main.cpp
@@ -170,6 +170,7 @@ int main(int argc, char* argv[])
 	settings.registerSetting(settings::ConnectionMatrix_ShowMediaLockedDot);
 	settings.registerSetting(settings::ConnectionMatrix_AllowCRFAudioConnection);
 	settings.registerSetting(settings::ConnectionMatrix_CollapsedByDefault);
+	settings.registerSetting(settings::ConnectionMatrix_ShowEntitySummary);
 
 	// Network
 	settings.registerSetting(settings::Network_ProtocolType);