+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - CTRL+SHIFT+L: Collapse Listeners
   - CTRL+T: Expand Talkers
   - CTRL+SHIFT+T: Collapse Talkers
-- Connection matrix displays Entity Connection Summary (squares) using color code (see Legend screen for more details)
+- Option to display Entity Connection Summary (squares) using color code (see Legend screen for more details)
 - Option to collapse all connection matrix headers for new entities
 - [Redundant stream name now based on underlying stream names, when possible](https://github.com/christophe-calmejane/Hive/issues/126)
 - [Popup message after critical error on the network interface](https://github.com/christophe-calmejane/Hive/issues/124)